Make window method faster (no add_const); fix dof
(make dof int in mkplot.)  Change from sympy solver to root finder of a
quadratic (meff-sol)^2.

+++ latfit/finalout/mkplot.py
@@ -215,10 +215,10 @@ def get_param_chisq(coords, dimops, result_min, fitrange=None):
     """Get chi^2 parameters."""
     param_chisq = namedtuple('param_chisq', ('redchisq', 'redchisq_round_str', 'dof'))
     if fitrange is None:
-        param_chisq.dof = len(coords)*dimops-len(result_min.x)
+        param_chisq.dof = int(len(coords)*dimops-len(result_min.x))
     else:
-        param_chisq.dof = (fitrange[1]-fitrange[0]+1)*dimops-len(result_min.x
-        )
+        param_chisq.dof = int((fitrange[1]-fitrange[0]+1)*dimops-len(
+            result_min.x))
     #Do this because C parameter is a fit parameter, it just happens to be guessed by hand
     if EFF_MASS and EFF_MASS_METHOD == 1 and C != 0.0:
         param_chisq.dof -= 1
